made society work

+++ server/society-management/sql.lua
@@ -1,4 +1,3 @@
-
 function society_management_callbacks_getSocieties()
     if ESX then
         -- get all societies from addon_account where shared = 1 and add money  to it from  addon_account_adata where name = @name from addon_account and the inventory from addon_incnentory and owner if one exists, and the
@@ -19,7 +18,6 @@ WHERE
 end
 
 function getSociety(jobName)
-
     if string.find(jobName, "society_") then
         jobname = jobName
     else
@@ -45,7 +43,7 @@ WHERE
 end
 
 function society_management_callback_deleteSociety(society)
-    if ESX then
+    if ESX and society then
         query = [[DELETE FROM addon_account WHERE name = @name;]]
         MySQL.query.await(query, { ['@name'] = society.name })
         query = [[DELETE FROM addon_account_data WHERE account_name = @name;]]
@@ -75,16 +73,21 @@ function society_managment_callback_editSociety(society)
 end
 
 function society_management_callback_withdrawSocietyMoney(society, amount, source)
-    if ESX  and society_management_callback_doesSocietyExist(society) then
-            society_management_framework_addMoney(source, amount)
+   
+    if ESX and society_management_callback_doesSocietyExist(society) and society_management_framework_hasEnoughMoney(source, amount) then
+        society_management_framework_addMoney(source, amount)
+    
+
         query = [[UPDATE addon_account_data SET money = money - @amount WHERE account_name = @name;]]
         MySQL.query.await(query, { ['@name'] = society.name, ['@amount'] = amount })
         return true
-        end
+    else
+        return false
+    end
 end
 
 function society_management_callback_depositSocietyMoney(society, amount, source)
-    if ESX  and society_management_callback_doesSocietyExist(society) and society_management_framework_hasEnoughMoney(source, amount) then
+    if ESX and society_management_callback_doesSocietyExist(society) and society_management_framework_hasEnoughMoney(source, amount) then
         society_management_framework_withdrawMoney(source, amount)
         query = [[UPDATE addon_account_data SET money = money + @amount WHERE account_name = @name;]]
         MySQL.query.await(query, { ['@name'] = society.name, ['@amount'] = amount })
@@ -92,19 +95,26 @@ function society_management_callback_depositSocietyMoney(society, amount, source
     else
         showNotification(source, Locale("not_enough_money"))
         return false
-end
+    end
 end
 
 function society_management_callback_setSocietyMoney(society, amount)
-    if ESX  and next(society)  and society_management_callback_doesSocietyExist(society?.name) then
-        query = [[UPDATE addon_account_data SET money = @amount WHERE account_name = @name;]]
-        MySQL.query.await(query, { ['@name'] = society.name, ['@amount'] = amount })
-        return true
+    if type(society) == "table" then
+        if ESX and next(society) and society_management_callback_doesSocietyExist(society.name) then
+            query = [[UPDATE addon_account_data SET money = @amount WHERE account_name = @name;]]
+            MySQL.query.await(query, { ['@name'] = society.name, ['@amount'] = amount })
+            return true
+        end
+    elseif type(society) == "string" then
+        if ESX and society_management_callback_doesSocietyExist(society) then
+            query = [[UPDATE addon_account_data SET money = @amount WHERE account_name = @name;]]
+            MySQL.query.await(query, { ['@name'] = society, ['@amount'] = amount })
+            return true
+        end
     end
 end
 
 function society_management_callback_doesSocietyExist(society)
-    
     if type(society) == "table" then
         societyName = society.name
     elseif type(society) == "string" then
